feat: el hashrate sale de la ecuacion y la frecuencia deja de arrastrar decimales
Dos arreglos que toca los mismos archivos, asi que van juntos.

1) EL HASHRATE FUERA

El fallo reportado: la estrategia PID no volvia a subir la frecuencia despues
de bajarla. La unica rama que la subia estaba dentro de un
`elif hashrate < setpoint`, asi que con el hashrate cumpliendo el objetivo el
tuner caia en un `else` que no proponia nada y los MHz que quitaba el calor no
volvian jamas. Y era el caso normal: se recomienda poner HASHRATE_SETPOINT por
debajo de la media medida.

El hashrate es un RESULTADO, no una palanca: no se puede fijar, solo observar.
Perseguirlo como setpoint hacia que el controlador reaccionara al ruido del pool
en vez de al estado del chip. Se quita del todo: fuera simple_pid y fuera los
objetos PID. `apply_strategy` sigue aceptando `hashrate=` por compatibilidad de
firma y lo descarta; tuning_manager ya no lo pasa. Se sigue midiendo y
escribiendo en el CSV, pero no decide nada.

tuning.py queda con cinco reglas por prioridad, un paso y una palanca por
muestra:

  1. temp > TARGET_TEMP        -> baja frecuencia (si no puede, voltaje)
  2. potencia sobre el limite  -> baja voltaje (si no puede, frecuencia)
  3. errores sobre el objetivo -> SUBE voltaje (si esta al maximo, baja freq)
  4. margen termico y freq por debajo del maximo -> SUBE frecuencia
  5. N muestras seguidas sin errores -> baja voltaje, buscando el minimo

La regla 5 espera `estable_para_bajar` muestras (toma ERROR_SETTLE) antes de
tocar nada: en el BM1370 de g el errorPercentage recorre casi 3 puntos con el
hardware quieto, asi que decidir con una sola muestra es perseguir ruido.

Tambien cae el segundo defecto del PID: `new_frequency = proposed_frequency`
adoptaba `Kp*error + integral`, que no es una frecuencia fisica, y desde
500 MHz podia proponer 400 de golpe.

2) LOS DECIMALES

En el CSV del nodo la frecuencia recorria 493.75, 498.75, 503.75... El valor con
coma entra de fuera (la web de AxeOS permite valores libres, y --frequency los
aceptaba), el tuner lo adoptaba tal cual y le sumaba pasos ENTEROS encima, asi
que el desfase no se corregia nunca. Ninguno de esos valores es uno que el
usuario haya configurado, y son mV y MHz que van a la API de un miner que no
aplica fracciones.

Se corta en los tres sitios: _a_rejilla() lleva al multiplo de FREQUENCY_STEP /
VOLTAGE_STEP mas cercano lo que se adopta de AxeOS, las dos estrategias
devuelven int en su recorte final (el unico punto por el que pasan todas las
salidas), y cli.py declara --voltage y --frequency como int para rechazar la
coma donde el usuario aun puede corregirla.

Verificado con el lazo cerrado contra un chip simulado: rampa hasta el borde del
objetivo de errores y quieto 55 muestras; 70C baja 850->775 y al enfriarse
recupera hasta 875; errores al 9% suben 1200->1220 mV y al calmarse baja de diez
en diez tras cada tanda de tres muestras estables; entrando con
1195.4mV/793.75MHz ninguna muestra posterior lleva decimales. 87 tests OK.

+++ cli.py
@@ -72,10 +72,13 @@ def parse_arguments() -> argparse.Namespace:
     parser.add_argument(
         "--fallback-stratum-user", type=str, help="Stratum user for backup pool"
     )
-    parser.add_argument("--voltage", type=float, help="Initial voltage override (mV)")
-    parser.add_argument(
-        "--frequency", type=float, help="Initial frequency override (MHz)"
-    )
+    # Enteros, no float: son mV y MHz que se escriben en el hardware y AxeOS no
+    # aplica fracciones. Aceptar "--frequency 493.75" solo conseguia meter un
+    # decimal que la estrategia arrastra sumandole pasos enteros el resto de la
+    # ejecucion (493.75, 498.75, 503.75...). Con type=int argparse lo rechaza en
+    # el momento, que es donde el usuario puede corregirlo.
+    parser.add_argument("--voltage", type=int, help="Initial voltage override (mV)")
+    parser.add_argument("--frequency", type=int, help="Initial frequency override (MHz)")
     parser.add_argument(
         "--sample-interval", type=float, help="Sample interval override (seconds)"
     )
